feat: Ensures that search results have the proper shelf value set.
This commit adds code that cross-references the search results with any books that are already stored in the local state and updates the shelf properties of the search results to match. Any search results that are not found in local state have their shelf value set to 'none'. This commit also adds support for keeping the shelf value in sync between the search results page and the list page.